MarioDraw: the swim base matrix is gated the other way and surfing reads mRoll as an f32
calcBaseMtx's wave-height branch runs *while* swimming (the ROM's beq), and
both surfing clamps read mPitch/mPitchMax and mRoll/mRollMax, never
mWaistRoll; mRoll is an f32 param, which its own -0.45f PARAM_INIT default
already said and which takes TSurfingParams::TSurfingParams to exact.
The triangle-normal branch crosses with the unnormalised up vector and
normalises that, not the binormal.

+++ include/Player/Mario.hpp
@@ -177,7 +177,9 @@ public:
 		TParamRT<f32> mWaistPitch;
 		TParamRT<s16> mWaistRollMax;
 		TParamRT<s16> mWaistPitchMax;
-		TParamRT<s32> mRoll;
+		// f32, not s32: its PARAM_INIT default is -0.45f and both
+		// TMario::calcBaseMtx branches read it with lfs.
+		TParamRT<f32> mRoll;
 		TParamRT<f32> mPitch;
 		TParamRT<s16> mRollMax;
 		TParamRT<s16> mPitchMax;
